ModifyTheAlphaOf_VS_VFX/requirements.txt
pywinauto~=0.6.8
vdf~=3.4
pillow~=9.5.0